Second iteration of performance tuning
Realised I should probably simplify this and base it around PK/FK constraints so as to be more robust in terms of which columns to choose for sorting and distributing. Also, makes tuning easier as user can now specify one ENV, POSTGRES_TO_REDSHIFT_OPTIMISED_FOR_TABLE, to select a table for which they'd like all queries to be optimised.