Update to aws-sdk v2
This removes the dependency on the json gem which is convenient because
the json gem just had a [security advisory][advisory] published against
it which we'd like to avoid.  We couldn't in our main app because it uses
this gem and this gem used aws-sdk-v1 which relied on an old json.

[advisory]: https://github.com/advisories/GHSA-jphg-qwrw-7w9g